Create implementation of deque
A deque (double-ended queue) is a data structure that allows elements to be inserted and removed from both the front and rear ends. It combines the properties of both stacks and queues.

Here are some important features of a deque:

Insertion and Deletion: You can insert and remove elements from both ends (front and rear).
Efficient Operations: Both operations at the front and rear have O(1) time complexity.
Flexible Usage: It can be used to implement both stacks (LIFO) and queues (FIFO) depending on the use case.
In your code, you're implementing a deque with methods such as:

pushback: Adds an element to the back.
pushfront: Adds an element to the front.
popback: Removes an element from the back.
popfront: Removes an element from the front.